Creadit management in generate endpoint

+++ backend/internal/db/user.go
@@ -77,7 +77,7 @@ func GetUserByID(id string, db *supa.Client) (*models.User, error) {
 }
 
 // UpdateUserCredits sets a user's credit balance to the provided value.
-func UpdateUserCredits(id string, credits int, db *supa.Client) (*models.User, error) {
+func UpdateUserCredits(id string, credits float64, db *supa.Client) (*models.User, error) {
 	updatedFields := map[string]interface{}{
 		"credits":    credits,
 		"updated_at": time.Now(),
