+++ backend/internal/models/user.go
@@ -7,7 +7,7 @@ type User struct {
 	Email            string    `json:"email"`
 	CreatedAt        time.Time `json:"created_at"`
 	PasswordHash     string    `json:"password_hash"`
-	Credits          int       `json:"credits"`
+	Credits          float64   `json:"credits"`
 	StripeCustomerID string    `json:"stripe_customer_id"`
 	Name             string    `json:"name"`
 	UpdatedAt        time.Time `json:"updated_at"`
